♻️ Extract Composer into dedicated file
Refactor holo-thread.tsx by extracting Composer and ComposerButton
into a new composer.tsx file.

- holo-thread.tsx: 2680 → 1783 lines (focused on message rendering)
- composer.tsx: 650 lines (input handling, voice, paste, animations)

The interface between them is minimal: one callback prop for marking
stopped messages. Both components access shared state via context hooks.
